apps/itti-backend/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "itti_backend" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/itti-backend/html'
 --cov-report xml:'../../coverage/apps/itti-backend/coverage.xml'
 --html='../../reports/apps/itti-backend/unittests/html/index.html'
 --junitxml='../../reports/apps/itti-backend/unittests/junit.xml'
"""

[project]
name = "itti-backend"
version = "1.0.0"
description = "FastAPI backend for ITTI technical test"
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.5.0",
  "langchain>=0.3.26",
  "langchain-openai>=0.3.27",
  "python-dotenv>=1.0.0",
  "textstat>=0.7.3",
  "pandas>=2.3.0",
  "sentence-transformers>=5.0.0",
  "scikit-learn>=1.6.1",
  "numpy>=2.0.2",
  "langchain-google-genai>=2.1.6",
]

[tool.hatch.build.targets.wheel]
packages = ["itti_backend"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
  "autopep8>=2.3.1",
  "ruff>=0.8.2",
  "pytest>=8.3.4",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []
